# cop-sys-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For COP-SYSTEM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/incident": {
            "get": {
                "tags": [
                    "Incident"
                ],
                "summary": "Get all incidents",
                "operationId": "1cdb6a26be5e4dd3aeed2eb623b0c65d",
                "responses": {
                    "200": {
                        "description": "Incidents list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/IncidentResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Incident"
                ],
                "summary": "Add a new incident",
                "operationId": "7cec8d625932531abb2156007c408a92",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateIncidentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Incident added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/incident/{incident}": {
            "get": {
                "tags": [
                    "Incident"
                ],
                "summary": "Get an incident by ID",
                "operationId": "b4da95be2ec9870c9133db534298c80a",
                "parameters": [
                    {
                        "name": "incident",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incident retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Incident not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Incident"
                ],
                "summary": "Update an existing incident",
                "operationId": "f8514609d3aaadca337e6399744f414d",
                "parameters": [
                    {
                        "name": "incident",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIncidentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Incident updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Incident not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Incident"
                ],
                "summary": "Remove an incident",
                "operationId": "7d7876343308995036e110c86ac12d20",
                "parameters": [
                    {
                        "name": "incident",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incident deleted successfully"
                    },
                    "404": {
                        "description": "Incident not found"
                    }
                }
            }
        },
        "/api/v1/location": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get all locations",
                "operationId": "27b3b420a9de6cd11a632f734f3c1303",
                "responses": {
                    "200": {
                        "description": "Locations list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LocationResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Add a new location",
                "operationId": "9e848f799e339bc39d28539c7d2b7fe7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLocationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/location/{location}": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get a location by ID",
                "operationId": "22319a6c9e43027abecd7f7d2b1acfce",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Location"
                ],
                "summary": "Update an existing location",
                "operationId": "62a803c10f5428e54260b5c057ede0af",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLocationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Location"
                ],
                "summary": "Remove a location",
                "operationId": "2800f6d1ab08b9ec3fa522fb5bd736bd",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location deleted successfully"
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateLocationRequest": {
                "required": [
                    "name",
                    "address",
                    "longitude",
                    "latitude",
                    "radius"
                ],
                "properties": {
                    "name": {
                        "type": "integer"
                    },
                    "address": {
                        "type": "string"
                    },
                    "longitude": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "string"
                    },
                    "radius": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "UpdateLocationRequest": {
                "properties": {
                    "name": {
                        "type": "integer"
                    },
                    "address": {
                        "type": "string"
                    },
                    "longitude": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "string"
                    },
                    "radius": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "LocationResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "integer"
                    },
                    "address": {
                        "type": "string"
                    },
                    "longitude": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "string"
                    },
                    "radius": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CreateIncidentRequest": {
                "required": [
                    "reporter_id"
                ],
                "properties": {
                    "incident_type": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "report_date_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "reporter_id": {
                        "type": "string"
                    },
                    "participants_id": {
                        "type": "string"
                    },
                    "vehicles_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateIncidentRequest": {
                "properties": {
                    "incident_type": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "report_date_time": {
                        "type": "string"
                    },
                    "reporter_id": {
                        "type": "string"
                    },
                    "participants_id": {
                        "type": "string"
                    },
                    "vehicles_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "IncidentResource": {
                "properties": {
                    "incident_id": {
                        "type": "integer"
                    },
                    "incident_type": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "report_date_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "reporter_id": {
                        "type": "string"
                    },
                    "participants_id": {
                        "type": "string"
                    },
                    "vehicles_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}